[script] fix wake up script for 0.2.1-1
Custom setups that don't have the bl_power for the touchscreen bl
were not able to wake up after sleep because of the assumption
that the bl_power file is there. This forces the pi to enable
display regardless if that file wasn't found.

+++ precompiled/opt/wake_up.sh
@@ -2,9 +2,14 @@
 
 # we need to probe the value of the bl_power first
 # to not wake up the screen if it's already up
-BL_PWR=`cat /sys/class/backlight/rpi_backlight/bl_power`
 
-if [ ${BL_PWR} -ne 0 ]; then
+if [ -f /sys/class/backlight/rpi_backlight/bl_power ]; then
+	BL_PWR=`cat /sys/class/backlight/rpi_backlight/bl_power`
+	
+	if [ ${BL_PWR} -ne 0 ]; then
+		vcgencmd display_power 1
+		echo 0 > /sys/class/backlight/rpi_backlight/bl_power
+	fi
+else
 	vcgencmd display_power 1
-	echo 0 > /sys/class/backlight/rpi_backlight/bl_power
 fi
